Fixed comments from Petr and Daniel
